+++ gamemodes/guesswho/entities/weapons/weapon_gw_decoy2.lua
@@ -1,3 +1,5 @@
+AddCSLuaFile()
+
 SWEP.Base = "weapon_gwbase"
 SWEP.Name = "Decoy 2.0"