+++ gamemodes/guesswho/entities/weapons/weapon_gw_default.lua
@@ -1,3 +1,5 @@
+AddCSLuaFile()
+
 SWEP.Base = "weapon_gwbase"
 SWEP.Name = "Default"