+++ gamemodes/guesswho/entities/weapons/weapon_gw_deflect.lua
@@ -1,3 +1,5 @@
+AddCSLuaFile()
+
 SWEP.Base = "weapon_gwbase"
 SWEP.Name = "Deflect"
 SWEP.AbilitySound = "ambient/energy/zap1.wav"