+++ gamemodes/guesswho/entities/weapons/weapon_gw_mind_control.lua
@@ -1,5 +1,7 @@
+AddCSLuaFile()
+
 SWEP.Base = "weapon_gwbase"
-SWEP.Name = "Mind Control"
+SWEP.Name = "Mind Transfer"
 
 SWEP.DrawGWCrossHair = true
 SWEP.AbilityDuration = 10
@@ -29,49 +31,9 @@ function SWEP:Ability()
 				fake:SetModel(oldModel)
 				fake:SetCollisionBounds( Vector(-8, - 8, 0), Vector(8, 8, 36) )
 			end)
-			self.remnant = fake
-			hook.Add("ScalePlayerDamage", "gw_ability_mind_control_prevent_death" .. ply:SteamID(), function(pl, hitgroup, dmginfo)
-				dmginfo:SetScale(0)
-			end)
-			timer.Create( "Ability.Effect.Mind.Control" .. ply:SteamID(), self.AbilityDuration, 1,
-				function()
-					if IsValid(self.remnant) then
-						local effect = EffectData()
-						effect:SetStart(ply:GetPos() + Vector(0, 0, 64))
-						effect:SetOrigin(ply:GetPos() + Vector(0, 0, 64))
-						effect:SetNormal(ply:GetAngles():Up())
-						util.Effect("ManhackSparks", effect, true, true)
-
-						local effect2 = EffectData()
-						effect2:SetOrigin(ply:GetPos() + Vector(0, 0, 64))
-						util.Effect("cball_explode", effect2)
-
-						ply:SetPos(self.remnant:GetPos())
-						ply:SetAngles(self.remnant:GetAngles())
-						ply:SetPlayerColor(self.remnant:GetPlayerColor())
-
-						local walker = ents.Create("npc_walker")
-						if !IsValid( walker ) then return end
-						walker:SetPos(self.remnant:GetPos())
-						walker:SetModel(self.remnant:GetModel())
-						walker:Spawn()
-						walker:SetAngles(self.remnant:GetAngles())
-						walker:Activate()
-						self.remnant:Remove()
-					end
-			end)
 		end
 	else
 		return true
-	end
-end
-
-function SWEP:OnRemove()
-	if SERVER then
-		local ply = self.Owner
-		timer.Remove( "Ability.Effect.Mind.Control" .. ply:SteamID() )
-		hook.Remove("ScalePlayerDamage", "gw_ability_mind_control_prevent_death" .. ply:SteamID())
-		if IsValid(self.remnant) then self.remnant:Remove() end
 	end
 end