+++ gamemodes/guesswho/entities/weapons/weapon_gw_seeker_crowbar.lua
@@ -1,3 +1,5 @@
+AddCSLuaFile()
+
 SWEP.PrintName = "Crowbar"
 
 if (CLIENT) then