+++ gamemodes/guesswho/entities/weapons/weapon_gw_smgdummy.lua
@@ -1,3 +1,5 @@
+AddCSLuaFile()
+
 SWEP.Base = "weapon_gw_default"
 
 SWEP.WorldModel = "models/weapons/w_smg1.mdl"