+++ gamemodes/guesswho/entities/weapons/weapon_gw_timelapse.lua
@@ -1,3 +1,5 @@
+AddCSLuaFile()
+
 SWEP.Base = "weapon_gwbase"
 SWEP.Name = "Timelapse"
 SWEP.AbilitySound = "gwabilities/timelapse.mp3"